テキストエリアをSectionTextAreaに切り出した
